+++ cherab/openadas/__init__.py
@@ -1,6 +1,6 @@
-# Copyright 2016-2018 Euratom
-# Copyright 2016-2018 United Kingdom Atomic Energy Authority
-# Copyright 2016-2018 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+# Copyright 2016-2022 Euratom
+# Copyright 2016-2022 United Kingdom Atomic Energy Authority
+# Copyright 2016-2022 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
 #
 # Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
 # European Commission - subsequent versions of the EUPL (the "Licence");
@@ -16,11 +16,8 @@
 # See the Licence for the specific language governing permissions and limitations
 # under the Licence.
 
-from .openadas import OpenADAS
+# for backward compatibility
+from ..atomic import AtomicData as OpenADAS
+
 from . import install
 from . import repository
-
-
-
-
-